fix in sync due to missing sid

+++ webnotes/sessions.py
@@ -46,7 +46,8 @@ def clear_cache(user=None):
 	webnotes.cache().delete_keys("session:")
 	if webnotes.session:
 		webnotes.cache().delete_keys("bootinfo:" + webnotes.session.user)
-		webnotes.cache().delete_keys("session:" + webnotes.session.sid)
+		if webnotes.session.sid:
+			webnotes.cache().delete_keys("session:" + webnotes.session.sid)
 	
 def clear_sessions(user=None, keep_current=False):
 	if not user:
